Added loop chunks of elements.
Most of the changes move 1-time init of data structures out
of the routines called in the loop (Grid constructor in particular).

Also began to remove threading, which is not going to be nessesary
in the future.  Reason: with chunked analysis, memory won't be
as much of a concern and the parallel map will represent sufficient
parallelization.