Added box counting for boxes that span multiple elements.
Grid.add() populates a boolean vector that keeps track of which
elements include the interface.  The vector is ordered such that
the first 8 elements are in a small box, the first 64 are in
a larger box, etc.  Outside of add, partial sums are taken
to note when the interface is in a larger box and to count
those boxes.

This seems to work, and doesn't add much more overhead than the
sub-element box counting.  It needs to be cleaned and tuned
before parallelizing.