First pass at improved fractal dimension algorithm.
The idea is to compute:
D - \log V(\epsilon) / \log(\epsilon)
where D is the dimension of the space (3 here), \epsilon is a
length, and V(\epsilon) is the volume of the union of
\epsilon-radius balls centered on every point in the interface.

We do this via Monte Carlo.  Random points are sampled in the
central 27-th of the volume.  For each, the shortest distance
to the interface is computed.  The histogram of those
distances is proportional to V(\epsilon).

The current impelmentation is very nearly brute force:
running shortest distances to a positive or negative field
point are kept for each sample.  Every element is checked to
see if it is within both running tallies.  If it is, then
every point in the element is checked to see if it is better.