Two changes: removed x,u transpose from nek.get_elem and added pad support.
NekFile.get_elem was transposing the vector fields (x,u).  This is a fine
thing to do, but doens't belong in the get_elem operation.  Removal required
some indexing changes through the rest of the code.

NekFile now looks for the '#max' prefix in the header and, if found, pads
the writes/reads to 8 MB blocks to match maxhutch/nek@d2d56b1 .